TourismManagementApp/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.0
cryptography==42.0.5
Django==5.0.3
django-ckeditor==6.7.1
django-filter==24.1
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.6
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.0
cryptography==42.0.5
datetime-truncate==1.1.1
Django==5.0.3
django-admin-charts==1.3.1
django-bower==5.2.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-cors-headers==4.4.0
django-filter==24.1
django-js-asset==2.2.0
django-light==0.1.0.post3
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.10.1
django-oauth-toolkit==2.3.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
Jinja2==3.1.4
jwcrypto==1.5.6
Markdown==3.6
MarkupSafe==2.1.5
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
pyngrok==7.1.6
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
